+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.10.3
 - Fixed typehints in Dataset
 - Dataset.create_train_test now takes a boolean `stratify` parameter.
+- Added default local filestorage when using `save_estimator`
 
 # v0.10.2
 - Fixed type inferences from data to sql in _load_data